Rename the orchestration layer to "server"
One name for the second distribution, everywhere it appears:

  orchestration/                            -> server/
  aind_behavior_vr_foraging_orchestration   -> aind_behavior_vr_foraging_server
  aind-behavior-vr-foraging-orchestration   -> aind-behavior-vr-foraging-server
  vr-foraging-orchestrator                  -> vr-foraging-server
  architecture/orchestration.md             -> architecture/server.md
  diagrams/orchestration.drawio             -> diagrams/server.drawio

Nothing about the design moved. The dependency still runs one way, server ->
packaging, still enforced by tests/test_package_boundary.py walking the AST of
every published module; the layer still owns the ledger, discovery, the stores,
the worker and the dashboard; and it is still `Private :: Do Not Upload`.

The ASCII diagram in server.md needed hand repair rather than substitution:
"server" is six characters narrower than "orchestrator", so every box the token
sat in lost its right border. The three affected rows are back to their original
widths, and the entrypoint name no longer wraps across two lines — it only did
so because "orchestrator" would not fit on one.

`docs/knowledge/log.md` keeps its history and gains an entry for this rename.
Identifiers and paths in the older entries were updated so their links and code
references still resolve — a change log pointing at files nobody can open is
worse than one whose prose is slightly anachronistic — and the new entry says so
explicitly, so no reader mistakes those entries for a package that was called
"server" at the time.

One incidental wording fix where the rename created a collision: server.md's
"no ledger server" (meaning no database service to run) became "no database
service", since in a document about a package named `server` the original could
only be read the wrong way.

Verified: 369 tests pass, ruff/ruff-format/ty/codespell clean, `uv lock`
regenerated (the old distribution is gone from the lockfile and the venv), the
image builds, and docker/smoke-test.sh passes against it end to end through the
renamed ENTRYPOINT.
