+++ docs/knowledge/architecture/index.md
@@ -18,12 +18,12 @@ Read in this order:
 - [batch.md](batch.md) — `process_sessions` / `aggregate`: **many** sessions into one queryable parquet export.
 - [cli.md](cli.md) — the `vr-foraging-packaging` command: one subcommand per pipeline function.
 - [data-contract-and-versioning.md](data-contract-and-versioning.md) — The `contraqctor` dataset, Harp streams, AIND metadata, and the three versions the code tracks.
-- [orchestration.md](orchestration.md) — the **second distribution**: how one session becomes one container, the `output.metadata.json` sidecar, what the work volume holds and for how long, and how to run the whole pipeline on a laptop.
+- [server.md](server.md) — the **second distribution**: how one session becomes one container, the `output.metadata.json` sidecar, what the work volume holds and for how long, and how to run the whole pipeline on a laptop.
 
 ## Repo layout
 
 Two distributions, one repo (a `uv` workspace). The dependency runs one way —
-orchestration → packaging — and `tests/test_package_boundary.py` enforces it.
+server → packaging — and `tests/test_package_boundary.py` enforces it.
 
 ```
 src/aind_behavior_vr_foraging_packaging/     # PUBLISHED to PyPI
@@ -51,13 +51,13 @@ src/aind_behavior_vr_foraging_packaging/     # PUBLISHED to PyPI
     ├── _session_metadata.py             # SessionMetadataProcessor (one row per session)
     └── _helper.py                       # slice_by_index, get_closest_from_timestamp
 
-orchestration/src/aind_behavior_vr_foraging_orchestration/   # NEVER published
+server/src/aind_behavior_vr_foraging_server/   # NEVER published
 ├── sidecar.py            # output.metadata.json + SidecarRecorder
 ├── process.py            # what runs inside a container: one session + its sidecar
 ├── ledger.py  models.py  # SQLite job queue
 ├── sources/  stores/     # which sessions exist / how bytes move
 ├── staging.py  runner.py  worker.py  dashboard.py
-└── cli.py                # `vr-foraging-orchestrator`
+└── cli.py                # `vr-foraging-server`
 ```
 
 Processor modules are private (`_`-prefixed); the public surface is
