+++ server/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=77", "setuptools-scm>=8"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "aind-behavior-vr-foraging-orchestration"
+name = "aind-behavior-vr-foraging-server"
 description = "Runs the VR Foraging packaging pipeline at scale: job ledger, session discovery, container worker, dashboard."
 authors = [
     {name = "Bruno Cruz", email = "bruno.cruz@alleninstitute.org"},
@@ -26,7 +26,7 @@ classifiers = [
 readme = {file = "README.md", content-type = "text/markdown"}
 
 dependencies = [
-    # The processor. This is the one direction the dependency runs: orchestration
+    # The processor. This is the one direction the dependency runs: server
     # knows about packaging, never the reverse — which is why the sidecar lives
     # here rather than in `pipeline/`. Resolved from the workspace root, not PyPI
     # (see [tool.uv.sources] below).
@@ -39,7 +39,7 @@ dependencies = [
 ]
 
 [project.scripts]
-vr-foraging-orchestrator = "aind_behavior_vr_foraging_orchestration.cli:main"
+vr-foraging-server = "aind_behavior_vr_foraging_server.cli:main"
 
 [tool.uv.sources]
 aind-behavior-vr-foraging-packaging = { workspace = true }
