+++ server/tests/test_ledger.py
@@ -3,8 +3,8 @@
 import threading
 
 import pytest
-from aind_behavior_vr_foraging_orchestration.ledger import Ledger, job_key
-from aind_behavior_vr_foraging_orchestration.models import Job
+from aind_behavior_vr_foraging_server.ledger import Ledger, job_key
+from aind_behavior_vr_foraging_server.models import Job
 
 
 def _make_ledger(tmp_path) -> Ledger:
