+++ tests/test_package_boundary.py
@@ -1,6 +1,6 @@
 """The published package must not depend on the unpublished one.
 
-`orchestration/` is a separate distribution that is never uploaded to PyPI. If
+`server/` is a separate distribution that is never uploaded to PyPI. If
 anything under `src/` imports it, the published wheel becomes installable but
 broken for everyone outside this repo — and the failure appears at *their* import
 time, not in our CI, which has both packages on the path.
@@ -15,7 +15,7 @@ from pathlib import Path
 import pytest
 
 _PUBLISHED = Path(__file__).resolve().parents[1] / "src" / "aind_behavior_vr_foraging_packaging"
-_FORBIDDEN = "aind_behavior_vr_foraging_orchestration"
+_FORBIDDEN = "aind_behavior_vr_foraging_server"
 
 
 def _imported_modules(source: str) -> set[str]:
@@ -35,10 +35,10 @@ def _imported_modules(source: str) -> set[str]:
     sorted(_PUBLISHED.rglob("*.py")),
     ids=lambda p: str(p.relative_to(_PUBLISHED)),
 )
-def test_published_package_does_not_import_the_orchestration_package(path: Path):
+def test_published_package_does_not_import_the_server_package(path: Path):
     offenders = {m for m in _imported_modules(path.read_text(encoding="utf-8")) if m.startswith(_FORBIDDEN)}
     assert not offenders, (
         f"{path.relative_to(_PUBLISHED)} imports {sorted(offenders)}. The dependency runs one way only: "
-        "orchestration -> packaging. Pass a callback in (see process_session's on_output/on_error) "
+        "server -> packaging. Pass a callback in (see process_session's on_output/on_error) "
         "rather than importing back out."
     )
